EES-645 - refactored manage content pages into the admin app, rearranged code to get initial page working
