+++ src/explore-education-statistics-admin/src/pages/release/edit-release/content/ReleaseContentPage.tsx
@@ -1,26 +1,26 @@
 import Link from '@admin/components/Link';
-import BasicReleaseSummary from '@admin/pages/release/edit-release/content/components/BasicReleaseSummary';
-import EditableTextRenderer from '@admin/pages/release/edit-release/content/components/EditableTextRenderer';
-import ManageReleaseContext, { ManageRelease } from '@admin/pages/release/ManageReleaseContext';
-import { getTimePeriodCoverageDateRangeStringShort } from '@admin/pages/release/util/releaseSummaryUtil';
+import BasicReleaseSummary from '@admin/modules/find-statistics/components/BasicReleaseSummary';
+import EditableTextRenderer from '@admin/modules/find-statistics/components/EditableTextRenderer';
+import ManageReleaseContext, {ManageRelease} from '@admin/pages/release/ManageReleaseContext';
+import {getTimePeriodCoverageDateRangeStringShort} from '@admin/pages/release/util/releaseSummaryUtil';
 import commonService from '@admin/services/common/service';
-import { IdTitlePair } from '@admin/services/common/types';
-import { Comment } from '@admin/services/dashboard/types';
+import {IdTitlePair} from '@admin/services/common/types';
+import {Comment} from '@admin/services/dashboard/types';
 import releaseContentService from '@admin/services/release/edit-release/content/service';
 import releaseSummaryService from '@admin/services/release/edit-release/summary/service';
-import { ReleaseSummaryDetails } from '@admin/services/release/types';
+import {ReleaseSummaryDetails} from '@admin/services/release/types';
 import Accordion from '@common/components/Accordion';
 import AccordionSection from '@common/components/AccordionSection';
 import FormFieldset from '@common/components/form/FormFieldset';
 import FormRadioGroup from '@common/components/form/FormRadioGroup';
 import WarningMessage from '@common/components/WarningMessage';
-import { AbstractRelease, ContentBlock, ReleaseType } from '@common/services/publicationService';
+import {AbstractRelease, ContentBlock, ReleaseType} from '@common/services/publicationService';
 import classNames from 'classnames';
-import React, { useContext, useEffect, useState } from 'react';
+import React, {useContext, useEffect, useState} from 'react';
 import PageSearchForm from '@common/components/PageSearchForm';
-import PublicationReleaseContent from '@common/modules/find-statistics/PublicationReleaseContent';
-import EditableMarkdownRenderer from './components/EditableMarkdownRenderer';
-import PrintThisPage from './components/PrintThisPage';
+import PublicationReleaseContent from "@admin/modules/find-statistics/PublicationReleaseContent";
+import EditableMarkdownRenderer from '../../../../modules/find-statistics/components/EditableMarkdownRenderer';
+import PrintThisPage from '../../../../modules/find-statistics/components/PrintThisPage';
 
 type PageMode = 'edit' | 'preview';
 
@@ -35,7 +35,7 @@ interface Model {
 const ReleaseContentPage = () => {
   const [model, setModel] = useState<Model>();
 
-  const { releaseId, publication } = useContext(
+  const {releaseId, publication} = useContext(
     ManageReleaseContext,
   ) as ManageRelease;
 
@@ -77,9 +77,7 @@ const ReleaseContentPage = () => {
           releases: [],
           legacyReleases: [],
           topic: {
-            theme: {
-              title: '',
-            },
+            theme
           },
           nextUpdate: '',
           contact: {
@@ -152,150 +150,13 @@ const ReleaseContentPage = () => {
             })}
           >
             <div className={model.pageMode === 'edit' ? 'page-editing' : ''}>
-              <h1 className="govuk-heading-l">
-                <span className="govuk-caption-l">
-                  {model.releaseSummary.timePeriodCoverage.label}{' '}
-                  {getTimePeriodCoverageDateRangeStringShort(
-                    model.releaseSummary.releaseName,
-                    '/',
-                  )}
-                </span>
-                {publication.title}
-              </h1>
-              <div className="govuk-grid-row">
-                <div className="govuk-grid-column-two-thirds">
-                  <div className="govuk-grid-row">
-                    <BasicReleaseSummary release={model.releaseSummary} />
-                  </div>
-                </div>
-              </div>
-              <hr />
-              <h2
-                className="govuk-heading-m govuk-!-margin-top-9"
-                data-testid="extra-information"
-              >
-                Help and support
-              </h2>
-              <Accordion
-                // publicationTitle={publication.title}
-                id="static-content-section"
-              >
-                <AccordionSection
-                  heading={`${publication.title}: methodology`}
-                  caption="Find out how and why we collect, process and publish these statistics"
-                  headingTag="h3"
-                >
-                  <p>
-                    Read our{' '}
-                    <Link to={`/methodology/${publication.methodologyId}`}>
-                      {`${publication.title}: methodology`}
-                    </Link>{' '}
-                    guidance.
-                  </p>
-                </AccordionSection>
-                {model.releaseSummary.type &&
-                model.releaseSummary.type.title ===
-                ReleaseType.NationalStatistics && (
-                  <AccordionSection
-                    heading="National Statistics"
-                    headingTag="h3"
-                  >
-                    <p className="govuk-body">
-                      The{' '}
-                      <a href="https://www.statisticsauthority.gov.uk/">
-                        United Kingdom Statistics Authority
-                      </a>{' '}
-                      designated these statistics as National Statistics in
-                      accordance with the{' '}
-                      <a href="https://www.legislation.gov.uk/ukpga/2007/18/contents">
-                        Statistics and Registration Service Act 2007
-                      </a>{' '}
-                      and signifying compliance with the Code of Practice for
-                      Statistics.
-                    </p>
-                    <p className="govuk-body">
-                      Designation signifying their compliance with the
-                      authority's{' '}
-                      <a href="https://www.statisticsauthority.gov.uk/code-of-practice/the-code/">
-                        Code of Practice for Statistics
-                      </a>{' '}
-                      which broadly means these statistics are:
-                    </p>
-                    <ul className="govuk-list govuk-list--bullet">
-                      <li>
-                        managed impartially and objectively in the public
-                        interest
-                      </li>
-                      <li>meet identified user needs</li>
-                      <li>produced according to sound methods</li>
-                      <li>well explained and readily accessible</li>
-                    </ul>
-                    <p className="govuk-body">
-                      Once designated as National Statistics it's a statutory
-                      requirement for statistics to follow and comply with the
-                      Code of Practice for Statistics to be observed.
-                    </p>
-                    <p className="govuk-body">
-                      Find out more about the standards we follow to produce
-                      these statistics through our{' '}
-                      <a href="https://www.gov.uk/government/publications/standards-for-official-statistics-published-by-the-department-for-education">
-                        Standards for official statistics published by DfE
-                      </a>{' '}
-                      guidance.
-                    </p>
-                  </AccordionSection>
-                )}
-                <AccordionSection heading="Contact us" headingTag="h3">
-                  <p>
-                    If you have a specific enquiry about {model.theme.title}{' '}
-                    statistics and data:
-                  </p>
-                  <h4 className="govuk-heading-s govuk-!-margin-bottom-0">
-                    {publication.contact && publication.contact.teamName}
-                  </h4>
-                  <p className="govuk-!-margin-top-0">
-                    Email <br />
-                    {publication.contact && (
-                      <a href={`mailto:${publication.contact.teamEmail}`}>
-                        {publication.contact.teamEmail}
-                      </a>
-                    )}
-                  </p>
-                  <p>
-                    {publication.contact && (
-                      <>
-                        Telephone: {publication.contact.contactName} <br />{' '}
-                        {publication.contact.contactTelNo}
-                      </>
-                    )}
-                  </p>
-                  <h4 className="govuk-heading-s govuk-!-margin-bottom-0">
-                    Press office
-                  </h4>
-                  <p className="govuk-!-margin-top-0">
-                    If you have a media enquiry:
-                  </p>
-                  <p>
-                    Telephone <br />
-                    020 7925 6789
-                  </p>
-                  <h4 className="govuk-heading-s govuk-!-margin-bottom-0">
-                    Public enquiries
-                  </h4>
-                  <p className="govuk-!-margin-top-0">
-                    If you have a general enquiry about the Department for
-                    Education (DfE) or education:
-                  </p>
-                  <p>
-                    Telephone <br />
-                    037 0000 2288
-                  </p>
-                </AccordionSection>
-              </Accordion>`
-            </div>
+
+
 
             <PublicationReleaseContent
-              data={model.release}
+              basicPublication={publication}
+              release={model.release}
+              releaseSummary={model.releaseSummary}
               styles={{}}
               TextRenderer={EditableTextRenderer}
               MarkdownRenderer={EditableMarkdownRenderer}
@@ -303,6 +164,7 @@ const ReleaseContentPage = () => {
               PrintThisPage={PrintThisPage}
               SearchForm={PageSearchForm}
             />
+            </div>
           </div>
         </>
       )}
